refactor(ast): change AstNodeKind::BinaryExpr to AstNodeKind::BinaryExpr(BinaryOp)
This commit changes the AstNodeKind::BinaryExpr to AstNodeKind::BinaryExpr(BinaryOp) to make it easier to access the operator of a binary expression. The operator is now stored in the AstNodeKind::BinaryExpr variant instead of a separate AstNodeKind::BinaryOp variant. The binary_operator method in SyntaxTreeBuilder has been removed since it is no longer needed.