feat(ast): add support for while statements
Add support for while statements in the AST by adding a new `WhileStmt` node to the `AstNodeKind` enum. Also, add a new `while_statement` method to the `SyntaxTreeBuilder` struct to create a new `WhileStmt` node. In addition, add a new `visit_while_statement` method to the `AstVisitor` trait and implement it in the `Evaluator` struct to evaluate while statements. Finally, add a new `while_stmt` method to the `Parser` struct to parse while statements.