feat(rlox): add support for detecting too many parameters and arguments
Add support for detecting too many parameters and arguments in the parser.
The maximum number of parameters and arguments is 255. When the limit is
exceeded, a `TooManyParameters` or `TooManyArguments` error is added to the
list of recovered errors.

Also, add two new test cases to check that the errors are correctly detected
and reported.