feat(rlox): add support for return statements without a value
This change adds support for return statements without a value in the
rlox language. The `return` keyword can now be used on its own to exit
a function and return `nil`. The `ret.lox` file in the `function`
directory of the `lox_files` directory has been added to test this
feature.

Also, the `Parser` struct in the `parser/mod.rs` file has been updated
to handle this new syntax. The `cur_token_on_same_line` method has been
added to check if the current token is on the same line as the previous
token. This is used to determine if the `return` statement has a value
or not.

Finally, the `test.rs` file in the `tests` directory has been updated
to include a new test case for the `ret.lox` file. Two new snapshot
files have also been added