feat(rlox): add support for returning from initializers
This commit adds support for returning from initializers. If an initializer
returns a value, it will raise a runtime error. The error message is
"RuntimeError: Cannot return from initializer". The error is raised when the
initializer returns a value other than nil.

A new test case is added to test this feature. The test case is named
"test__runtime_errors__init_ret". It tests that a runtime error is raised when
an initializer returns a value other than nil. The test case passes if the
expected error message is printed.

The following files are modified:
- rlox/lox_files/runtime_errors/init_ret.lox
- rlox/src/interpreter/diagnostics.rs
- rlox/src/interpreter/mod.rs
- rlox/src/interpreter/types.rs
- rlox/tests/snapshots/test__runtime_errors__init_ret.snap
- rlox/tests/test.rs